Extracted some of the views of MovieDetailView the into their own separate subviews.  Added a ProgressView and necessary logic.
